Course 4 - Data Pipeline/airflow_sparkify_project/requirements.txt
apache-airflow
apache-airflow-providers-amazon
psycopg2-binary
boto3
python-dotenv
SQLAlchemy
